+++ CHANGELOG.md
@@ -6,7 +6,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## 0.3
 
-- Support for `piral.json` and `pilet.json` (#28)
+- Added support for `piral.json` and `pilet.json` (#28)
+- Added support for Bun as package manager
+- Fixed issue if `dependencies` section of *package.json* does not exist
 - Updated dependencies
 
 ## 0.2